% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullview.R
\name{fullview}
\alias{fullview}
\title{View underlying simulation details for \code{TriLLIEM.sim} objects}
\usage{
fullview(sim)
}
\arguments{
\item{sim}{an object of class \code{TriLLIEM.sim}}
}
\value{
A data frame with the same counts in each category as \code{sim}, but
with underlying simulation probabilities given as columns:
\describe{
\item{"\code{typeOrig}"}{Index for each genetically distinct row,
based on "\code{M}", "\code{F}", "\code{C}", "\code{matOrg}", and "\code{patOrg}".
Ranges from 1 to 64, unlike the \code{type} column in \code{sim} which ranges from 1
to 60, as the inclusion of "\code{matOrg}" and "\code{patOrg}" lead to potentially
four additional rows.}
\item{"\code{matOrg}" and "\code{patOrg}"}{Binary variables for when the minor allele
is maternally/paternally inherited, so that counts when "\code{M}", "\code{F}", "\code{C}"
are all 1 are not ambiguous when determining parent-of-origin effects.}
\item{"\code{prMF}"}{Probability of the mother ("\code{M}") and father ("\code{F}") pairing
in the simulated population, based on the minor allele frequency and mating
type coefficients given during simulation.}
\item{"\code{prCGivenMFOrg}"}{Probability of the child's genotype ("\code{C}") and the
allele parent of origin ("\code{matOrg}" and "\code{patOrg}") conditional on the
genotypes of the mother and father.}
\item{"\code{prMFCOrg}"}{Probability of the triad based on genotypes and parent
of origin, product of "\code{prMF}" and "\code{prCGivenMFOrg}".}
\item{"\code{PrMFCOrg}"}{Probability of the triad based on genotypes, parent of
origin, and environmental exposure ("\code{E}") conditional on disease status
("\code{D}"), obtained by scaling "\code{prMFCOrg}" by the risk of disease (equal to
"\code{prMFCOrg}" when "\code{D}" is 0).}
\item{"\code{pop}"}{Sub-population each row is simulated from (all \code{1} if
simulated without population stratification)}
}

The "\code{count}" column in "\code{sim}" is the sum of the rows of the "\code{count}"
column in the returned data frame when grouped by "\code{M}", "\code{F}", "\code{C}", "\code{E}",
"\code{D}".
}
\description{
View underlying simulation details for \code{TriLLIEM.sim} objects
}
\examples{

## View the underlying distributions behind some of the example models given in simulateData()
dat1 <- simulateData(S = c(1, 2, 4), If = 3)
fullview(dat1)

dat2 <- simulateData(
  nControl = 1000,
  propE = c(0.1, 0.4),
  propE.control = c(0.2, 0.2),
  nPop = 2,
  maf = c(0.3, 0.4),
  prev.byPop = c(0.2, 0.3),
  prop.byPop = c(0.6, 0.4)
 )
fullview(dat2)
}
