% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.TriLLIEM.R
\name{nobs.TriLLIEM}
\alias{nobs.TriLLIEM}
\title{Accessing number of observations in TriLLIEM objects}
\usage{
\method{nobs}{TriLLIEM}(object)
}
\arguments{
\item{object}{a \code{TriLLIEM} object}
}
\value{
an integer representing the number of observations in the data used
to get \code{object}.
}
\description{
Accessing number of observations in TriLLIEM objects
}
\examples{
model <- TriLLIEM(dat = example_dat4R)
nobs(model)
}
