% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.TriLLIEM.R
\name{residuals.TriLLIEM}
\alias{residuals.TriLLIEM}
\title{Accessing residuals for TriLLIEM objects}
\usage{
\method{residuals}{TriLLIEM}(
  object,
  type = c("deviance", "pearson", "working", "response", "partial"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Accessing residuals for TriLLIEM objects
}
