% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TriLLIEM.R
\name{TriLLIEM}
\alias{TriLLIEM}
\title{Running a loglinear analysis of trio data.}
\usage{
TriLLIEM(
  mtmodel = "MS",
  effects = c("C", "M"),
  dat,
  PStest = FALSE,
  includeE = FALSE,
  Einteraction = "M",
  includeD = FALSE,
  includeIm = FALSE,
  includeIf = FALSE,
  Minit = 0.5,
  max.iter = 30,
  EM.diag = FALSE
)
}
\arguments{
\item{mtmodel}{Mating type model to use in the analysis, can be "\code{HWE}" for
Hardy-Weinberg Equilibrium, "\code{MS}" for Mating Symmetry, and "\code{MaS}" for Mating Asymmetry.}

\item{effects}{A vector listing the effects, as strings, to include
in the model.  Example effects include:
\describe{
\item{"\code{C}"}{Child effects.}
\item{"\code{M}"}{Maternal effects.}
\item{"\code{Im}"}{Maternal imprinting effects.}
\item{"\code{If}"}{Paternal imprinting effects.}
}}

\item{dat}{A data frame with triad data, with the formatting of
\link{example_dat4R}.}

\item{PStest}{A logical value indicating whether to perform a population
stratification test on the data.}

\item{includeE}{A logical value indicating whether to include environment
interaction effects.}

\item{Einteraction}{A string indicating what variable environmental effects
interact with.  Can be "\code{Im}", "\code{If}", "\code{C}", or "\code{M}".}

\item{includeD}{A logical value indicating whether to use the hybrid
model with controls.}

\item{includeIm}{A logical value indicating whether to include maternal
imprinting in the model, equivalent to adding "\code{Im}" in the "\code{effects}" vector.}

\item{includeIf}{A logical value indicating whether to include paternal
imprinting in the model, equivalent to adding "\code{If}" in the "\code{effects}" vector.}

\item{Minit}{Initial proportion of maternal inheritence to split the triple
heterozygote cell by if the EM algorithm is necessary.}

\item{max.iter}{Maximum number of iterations for the EM algorithm.}

\item{EM.diag}{A logical value indicating whether to show diagnostic messages
for the EM algorithm.}
}
\value{
An object of class "\code{glm}".
}
\description{
Running a loglinear analysis of trio data.
}
\examples{
res <- TriLLIEM(mtmodel = "HWE", effects = c("C", "M", "Im"), dat = example_dat4R)
summ_trill(res)
}
