% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.TriLLIEM.R
\name{vcov.TriLLIEM}
\alias{vcov.TriLLIEM}
\title{Calculate Variance-Covariance Matrix for a \code{TriLLIEM} Object}
\usage{
\method{vcov}{TriLLIEM}(object, ...)
}
\arguments{
\item{object}{a \code{TriLLIEM} object.}

\item{...}{additional arguments, UNUSED.}
}
\value{
A matrix of covariances between parameters, accounting for the EM
algorithm.
}
\description{
Calculate Variance-Covariance Matrix for a \code{TriLLIEM} Object
}
\examples{
model <- TriLLIEM(dat = example_dat4R, effects = c("C", "M", "Im"))
vcov(model)
}
\keyword{internal}
