% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.TriLLIEM.R
\name{anova.TriLLIEM}
\alias{anova.TriLLIEM}
\title{Analysis of Deviance for TriLLIEM objects}
\usage{
\method{anova}{TriLLIEM}(object, ..., dispersion = NULL, test = NULL)
}
\arguments{
\item{test}{}
}
\description{
Analysis of Deviance for TriLLIEM objects
}
