% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dat4R}
\alias{example_dat4R}
\title{Example triad data}
\format{
A dataframe with columns for:
\describe{
\item{type}{index for the category corresponding to maternal (M), paternal
(F), and child (C) genotypes}
\item{mt_MS}{mating type category for maternal (M) and paternal (F)
genotypes under a mating symmetry model}
\item{mt_MaS}{mating type category for maternal (M) and paternal (F)
genotypes under a model that does not assume mating symmetry}
\item{M}{mother genotype}
\item{F}{father genotype}
\item{C}{child genotype}
\item{E}{binary variable for if environmental effects are present (1) or
not present (0)}
\item{D}{disease present (1)/not present (0)}
\item{count}{counts for each category of maternal (M), paternal (F), and
child (C) genotypes}
}
}
\usage{
example_dat4R
}
\description{
Example data formatted to be used in the \link{TriLLIEM} function.
}
\keyword{datasets}
