% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.TriLLIEM.R
\name{nobs.TriLLIEM}
\alias{nobs.TriLLIEM}
\title{Accessing number of observations in TriLLIEM objects}
\usage{
\method{nobs}{TriLLIEM}(object, ...)
}
\arguments{
\item{...}{}
}
\description{
Accessing number of observations in TriLLIEM objects
}
