% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.TriLLIEM.R
\name{print.summary.TriLLIEM}
\alias{print.summary.TriLLIEM}
\title{Title}
\usage{
\method{print}{summary.TriLLIEM}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"),
  show.residuals = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
