% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.TriLLIEM.R
\name{print.TriLLIEM}
\alias{print.TriLLIEM}
\title{Print method for \code{TriLLIEM} objects}
\usage{
\method{print}{TriLLIEM}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \code{TriLLIEM}.}

\item{digits}{number of digits to display.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints details of the \code{TriLLIEM} model, only including genetically
relevant parameters in listed coefficients.  To view all fitted parameters,
run \code{coef(x)}.
}
\description{
Print method for \code{TriLLIEM} objects
}
