% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEMIM.R
\name{runEMIM}
\alias{runEMIM}
\title{Title}
\usage{
runEMIM(
  mtmodel = "MS",
  effects = c("C", "M"),
  peddat,
  emimpath = "~/EMIM/",
  Einteraction = "M",
  includeIm = FALSE,
  includeIf = FALSE,
  weinberg = FALSE,
  includeE = FALSE
)
}
\arguments{
\item{weinberg}{}
}
\description{
Title
}
\keyword{internal}
