% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHaplin.R
\name{runHaplin}
\alias{runHaplin}
\title{Run Haplin on a data set}
\usage{
runHaplin(
  effects = c("C", "M"),
  dat,
  includeD = FALSE,
  includeE = FALSE,
  PoO = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Run Haplin on a data set
}
\keyword{internal}
