% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateData}
\alias{simulateData}
\title{Simulate data for TriLLIEM}
\usage{
simulateData(
  ntrios = 1000,
  maf = 0.3,
  R = c(1, 1, 1),
  S = c(1, 1, 1),
  V = c(1, 1, 1),
  mtCoef = c(1, 1, 1),
  Im = 1,
  If = 1,
  includeE = FALSE,
  Einteraction = "M",
  propE = 0,
  includeControl = FALSE,
  nControl = 0,
  propE.control = propE,
  includePopStrat = FALSE,
  numPop = 1,
  Fst = 0.005,
  prev.byPop = NULL,
  prop.byPop = NULL
)
}
\arguments{
\item{ntrios}{Number of trios to simulate.}

\item{maf}{Minor allele frequency in the population, a proportion between 0 and 1.}

\item{R}{Vector of 3 elements representing child effects for 0, 1, and 2
copies of the risk allele, respectively.}

\item{S}{Vector of 3 elements representing maternal effects for 0, 1, and 2
copies of the risk allele, respectively.}

\item{V}{Vector of 3 elements representing gene-environment effects
for 0, 1, and 2 copies of the risk allele, respectively.}

\item{mtCoef}{Mating type coefficients.}

\item{Im}{Maternal imprinting effect.}

\item{If}{Paternal imprinting effect.}

\item{includeE}{A logical value indicating whether environmental effects
should be included in the simulation.}

\item{Einteraction}{A string indicating what variable environmental effects
interact with.  Can be "\code{Im}", "\code{If}", "\code{C}", or "\code{M}".}

\item{propE}{Proportion of case trios in the environmental exposure group by
population, assumed equal to case population proportion.}

\item{includeControl}{A logical value indicating whether controls should
be included in the simulations.}

\item{nControl}{Number of control trios, must be less than ntrios.}

\item{propE.control}{Proportion of control trios in the environmental exposure
group by population, assumed equal to case population proportion.}

\item{includePopStrat}{A logical value indicating whether to include
population stratification in the simulation.}

\item{numPop}{not currently used.}

\item{Fst}{not currently used.}

\item{prev.byPop}{Prevalence of cases in each sub population.}

\item{prop.byPop}{Proportion of each sub population, must sum to 1.}
}
\value{
A data frame of the same format as \link{example_dat4R}
}
\description{
Simulate data for TriLLIEM
}
\examples{
## Simulating data with multiplicative maternal effect of 2, and paternal imprinting of 3.
simulateData(S = c(1, 2, 4), If = 3)
}
