% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.TriLLIEM.R
\name{summary.TriLLIEM}
\alias{summary.TriLLIEM}
\title{Summary function for TriLLIEM functions}
\usage{
\method{summary}{TriLLIEM}(object)
}
\arguments{
\item{object}{an object of class "\code{TriLLIEM}, usually, a result of a call to
\link{TriLLIEM}.(FORMATTING HERE
WAS WORD FOR WORD TAKEN DIRECTLY FROM THE DOCUMENTATION OF ?summary.glm
, MAKING NOTE OF THIS IN CASE THIS NEEDS TO BE CITED SOMEHOW)}
}
\value{
A list with the same components as those returned by \link{summary.glm},
but with the addition of the following:
\describe{
\item{terms}{the component from \code{object}.}
\item{aic}{the component from \code{object}.}
\item{EM_iter}{the component from \code{object}}
}
}
\description{
Summary function for TriLLIEM functions
}
\details{
Due to \link{TriLLIEM} using the EM algorithm when fitting imprinting effects,
calculation of residuals is modified compared to in \link{summary.glm}, where
the original data (\code{object$y_initial}) is used instead of \code{object$y} to
ensure the correct residuals are computed.  See \link{TriLLIEM} for more details.
}
\examples{
res <- TriLLIEM(mtmodel = "HWE", effects = c("C", "M", "Im"), dat = example_dat4R)
summary(res)
}
