% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.TriLLIEM.R
\name{summary.TriLLIEM}
\alias{summary.TriLLIEM}
\title{Summary function for TriLLIEM functions}
\usage{
\method{summary}{TriLLIEM}(object, ...)
}
\arguments{
\item{object}{an object of class "\code{TriLLIEM}", resulting from a call to
\link{TriLLIEM}.}

\item{...}{arguments passed to or from other methods.}
}
\value{
A list with the same components as those returned by \link{summary.glm},
but with the addition of the following:
\describe{
\item{terms}{the same component from \code{object}, always included for "\code{TriLLIEM}" objects.}
\item{aic}{the same component from \code{object}, always included for "\code{TriLLIEM}" objects}
\item{EM_iter}{the same component from \code{object}, always included for "\code{TriLLIEM}" objects}
}
}
\description{
Summary function for TriLLIEM functions
}
\details{
Due to \link{TriLLIEM} using the EM algorithm when fitting imprinting effects,
calculation of residuals is modified compared to in \link{summary.glm}, where
the original data (\code{object$y_initial}) is used instead of \code{object$y} to
ensure the correct residuals are computed.  See \link{TriLLIEM} for more details.
}
\examples{
res <- TriLLIEM(mtmodel = "HWE", effects = c("C", "M", "Im"), dat = example_dat4R)
summary(res)

}
