% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.TriLLIEM.R
\name{summary.TriLLIEM}
\alias{summary.TriLLIEM}
\title{Summary function for TriLLIEM functions}
\usage{
\method{summary}{TriLLIEM}(object)
}
\arguments{
\item{res}{Resulting object of class "\code{glm} from the \code{\link[=TriLLIEM]{TriLLIEM()}} function.}
}
\value{
A list with the following components:
\describe{
\item{effects}{The maximum likelihood estimates of the coefficients in the model.}
\item{se}{The standard errors of the maximum likelihood estimates, based on the observed information matrix.}
\item{pvals}{The p-values for each of the maximum likelihood estimates.}}
}
\description{
Summary function for TriLLIEM functions
}
\examples{
res <- TriLLIEM(dat = example_dat4R)
summ_trill(res)
}
